# backend/app/static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is the Backend API for the Task Management System serving the Vue.js Frontend",
    "version": "1.0.0",
    "title": "Task Management System - Flask App API",
    "contact": {
      "email": "sandeepbl@gmail.com"
    }
  },
  "host": "localhost:5000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/users/{id}/": {
      "get": {
        "summary": "View User Profile",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "User profile"
          }
        }
      },
      "put": {
        "summary": "Modify User",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "User modified"
          }
        }
      },
      "delete": {
        "summary": "Delete User",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "User deleted"
          }
        }
      }
    },
...
  }
}